Move internal type to an internal package
